Generate baker keys
